Introduced the eager ID3 de-unsychronization filter with a minimal (not working) implementation. This eager implementation will try to convert all data before any reading can be performed. In the next commits, every addition to the filter or reader class needs VERY specific rational.
This just adds the appropriate core for the filter to the Reader class, and outlines the possible usage for the filter class.
"make check" returns no errors, as reading unsynchronized data is not yet part of the test suite.
Run "./id3inspector examples/008_id3_2.3_MCDI_unsynchronized" to see, where the current implementation failes.